% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_additionalColumns.R
\name{additionalColumnsUI_main}
\alias{additionalColumnsUI_main}
\alias{additionalColumnsUI_sidebar}
\alias{additionalColumns}
\title{Additional Columns module}
\usage{
additionalColumnsUI_main(id, style = "font-size:80\%")

additionalColumnsUI_sidebar(id)

additionalColumns(id, Table, defaultCols = NULL, maxAdditionalCols = 5, ...)
}
\arguments{
\item{id}{shiny id}

\item{Table}{}

\item{defaultCols}{}

\item{maxAdditionalCols}{}

\item{...}{additional arguments passed to datatable}
}
\description{
Additional Columns module
}
