% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addTumortypes}
\alias{addTumortypes}
\title{Add tumor types}
\usage{
addTumortypes(data, anno)
}
\arguments{
\item{data}{data.frame, `celllinename` column required}

\item{anno}{data.frame, cell line annotation}
}
\value{
data data.frame with the additional column
}
\description{
A simple function that adds `tumortype` column to the data
}
