% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_brushPlot.R
\name{brushPlotUI}
\alias{brushPlotUI}
\alias{brushPlot}
\title{Brush Plot Module}
\usage{
brushPlotUI(id, ..., direction = "x", height = "600px")

brushPlot(
  input,
  output,
  session,
  plotExpr,
  checkExpr,
  textCallback = defaultTextCallback,
  defaultCutoff_x = NULL,
  step_x = 0.1,
  dx = 0.5,
  sorted_x = "no",
  defaultCutoff_y = -1,
  step_y = 0.1,
  dy = 0,
  sorted_y = "desc",
  availableChoices = c("cutoff", "number", "percentile"),
  test = FALSE,
  ...
)
}
\arguments{
\item{...}{}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{plotExpr}{}

\item{checkExpr}{}

\item{textCallback}{}

\item{defaultCutoff_x}{}

\item{step_x}{}

\item{dx}{}

\item{sorted_x}{}

\item{defaultCutoff_y}{}

\item{step_y}{}

\item{dy}{}

\item{sorted_y}{}

\item{availableChoices}{}

\item{test}{}
}
\value{

}
\description{
Brush Plot Module
}
