% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{buildMachineLearning}
\alias{buildMachineLearning}
\title{Workhorse for machine learning.}
\usage{
buildMachineLearning(
  cs,
  geneSet,
  geneAnno,
  species = "human",
  classColumn = "class",
  classLabel = NULL,
  p_test = 0,
  itemsColumn = getOption("xiff.column"),
  trainingData = NULL,
  getDataForModelFnc = getDataForModel,
  dataParams = NULL,
  method = "rf",
  tuneLength = 5,
  number = 10,
  repeats = 10,
  selectBestFeaturesFnc = "auto",
  threshold = "Confirmed",
  maxFeatures = "auto",
  featuresParams = NULL,
  ...,
  .verbose = TRUE,
  task = FALSE,
  .epsilonRNAseq = 10,
  .otherParams = list(),
  .extraClass = NULL
)
}
\arguments{
\item{cs}{class selection. Can be a list with two character vectors or
data.frame.}

\item{geneSet}{list of genes to be used as model features.}

\item{geneAnno}{gene annotation table.}

\item{species}{species.}

\item{classColumn}{name of the column that will be used to assgin classes.
Used when cs is data.frame.}

\item{classLabel}{classLable assigment (list with class1_name and
class2_name) or single string containing the name of the positive
class.}

\item{p_test}{percentage of the cs that will be assigned as test.}

\item{itemsColumn}{if cs is data.frame, specifies the name of column that 
contains items (celllines or tissues). Usually it should not be
used.}

\item{trainingData}{data.frame if user wants to use custom data instead of 
the result of \code{getDataForModel}.}

\item{getDataForModelFnc}{if the user wants to use custom function to get 
the data it can repleace \code{getDataForModel}.}

\item{dataParams}{list with additional params to be passed to
\code{getDataForModelFnc}}

\item{method}{name of ml methods (use values from 
\code{xiffSupportedModels()}) or model names from caret package.}

\item{tuneLength}{caret fitting parameter.}

\item{number}{caret fitting parameter.}

\item{repeats}{caret fitting parameter.}

\item{selectBestFeaturesFnc}{}

\item{threshold}{threshold for feature selection.}

\item{maxFeatures}{max number of features to be used in the fitting process.
Each feature selection methods contain a way to limit its number.}

\item{featuresParams}{additional params to be passed to selectBestFeaturesFnc
if custom function is used.}

\item{...}{}

\item{.verbose}{logical. if true prints additional values.}

\item{task}{internal param to be used by applications.}

\item{.epsilonRNAseq}{GREP param.}

\item{.otherParams}{list of other parameters to be saved with the model.}

\item{.extraClass}{other class to be added to \code{class} vector. Allows to
use custom methods. For advanced users.}
}
\value{
MLXIFF object
}
\description{
Workhorse for machine learning.
}
