% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAssigment-helpers.R
\name{getClassLabel}
\alias{getClassLabel}
\alias{getClassLabelVector}
\alias{unlistClassAssignment}
\alias{addClassAssigmentAttribute}
\alias{getClassAssigmentAttribute}
\title{Get class label list from classAssignment object.}
\usage{
getClassLabel(ca)

getClassLabelVector(ca)

unlistClassAssignment(ca)

addClassAssigmentAttribute(x, ca)

getClassAssigmentAttribute(x)
}
\arguments{
\item{ca}{classAssigment object}

\item{x}{an ClassAssigmentInAttribute object or object to be transformed to 
ClassAssigmentInAttribute.}
}
\description{
Get class label list from classAssignment object.

Utility function for working with classAssignment objects.
}
\details{
This function is dedicated mostly for CLIFF/XIFF developers.
}
\examples{

ca <- classAssignment(cl1 = c("setosa"), cl2 = c("versicolor", "virginica"))
iris2 <- addClassAssigmentAttribute(iris, ca)

getClassLabel(ca)
getClassLabel(iris2)
getClassLabelVector(ca)
getClassLabelVector(iris2)

}
