% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{classAssignment}
\alias{classAssignment}
\title{Class assignment}
\usage{
classAssignment(...)
}
\arguments{
\item{...}{named character vectors}
}
\value{
classAssignment object
}
\description{
Helper class that holds class samples with class labels
}
