% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utl_classDetails.R
\name{classDetailsWrapperUI_main}
\alias{classDetailsWrapperUI_main}
\alias{classDetailsWrapperUI_show}
\alias{classDetailsWrapper}
\title{Class Details Shiny Module}
\usage{
classDetailsWrapperUI_main(
  id,
  defaults = c("sensitive", "resistant"),
  inline = FALSE,
  allowNameChange = TRUE
)

classDetailsWrapperUI_show(id)

classDetailsWrapper(
  input,
  output,
  session,
  classLabel,
  classSelection,
  Selected,
  onChange
)
}
\arguments{
\item{inline}{logical, widget orientation}

\item{allowNameChange}{logical}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{classLabel}{class labels reactive}

\item{classSelection}{class selection reactive}

\item{Selected}{selected items}

\item{onChange}{function called on change}
}
\description{
Class Details Shiny Module
}
