% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{classIdToLabel}
\alias{classIdToLabel}
\title{Convert classId to class labels}
\usage{
classIdToLabel(x, ca)
}
\arguments{
\item{x}{vector with class labels (class1/class2)}

\item{ca}{classAssignment object}
}
\value{
vector containing the labels from classAssignment
}
\description{
Convert classId to class labels
}
\examples{

ca <- classAssignment(sensitive = c("a", "b", "c"), resistant = NULL)
classIdToLabel(c("class1", "class2", "class2", "class1", "class1"), ca)

}
