% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{classLabel2levels}
\alias{classLabel2levels}
\title{Convert classLabel list into vector of levels}
\usage{
classLabel2levels(cl)
}
\arguments{
\item{cl}{classLabel list}
}
\value{
character vector with 2 elements.
}
\description{
Convert classLabel list into vector of levels
}
\examples{

cl1 <- list(class1_name = "cl1", class2_name = "cl2")
classLabel2levels(cl1)

cl2 <- list(class2_name = "cl2", class1_name = "cl1")
classLabel2levels(cl1) == classLabel2levels(cl2)

}
