% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{column_1}
\alias{column_1}
\alias{column_2}
\alias{column_3}
\alias{column_4}
\alias{column_5}
\alias{column_6}
\alias{column_7}
\alias{column_8}
\alias{column_9}
\alias{column_10}
\alias{column_11}
\alias{column_12}
\alias{fluidRow_12}
\title{Convenience Shiny Column Functions}
\usage{
column_1(...)

column_2(...)

column_3(...)

column_4(...)

column_5(...)

column_6(...)

column_7(...)

column_8(...)

column_9(...)

column_10(...)

column_11(...)

column_12(...)

fluidRow_12(...)
}
\arguments{
\item{...}{shiny ui elements}
}
\description{
Convenience Shiny Column Functions
}
