% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{commonPlotTheme}
\alias{commonPlotTheme}
\title{XIFF ggplot2 theme}
\usage{
commonPlotTheme(legend.position = "right", textSize = 16)
}
\arguments{
\item{legend.position}{where the legend need to be located}

\item{textSize}{base text size}
}
\value{

}
\description{
XIFF ggplot2 theme
}
\examples{

df <- data.frame(
  gp = factor(rep(letters[1:3], each = 10)),
  y = rnorm(30)
)
ggplot(df, aes(gp, y)) + geom_point() + commonPlotTheme()


}
