% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{containerDT}
\alias{containerDT}
\title{DT Container with smaller font}
\usage{
containerDT(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
div with datatable container
}
\description{
DT Container with smaller font
}
