% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-params.R
\name{continuousTestByVariable}
\alias{continuousTestByVariable}
\title{Continuous Test}
\usage{
continuousTestByVariable(
  data,
  byVar,
  x,
  y,
  test = matrixTests::row_cor_pearson,
  ...
)
}
\arguments{
\item{data}{data.frame containing \code{byVar}, \code{x}, \code{y} columns.}

\item{byVar}{a column name used to split the data for tests.}

\item{x}{column name from \code{data}, which contains continuous variable.}

\item{y}{column name from \code{data}, which contains continuous variable.}

\item{test}{function used to perform the tests. It needs to take two matrices.}

\item{...}{other params passed to test function.}
}
\value{
data.frame with statistical tests applied across all the \code{byVar} levels.
}
\description{
Continuous Test
}
\examples{

set.seed(123)

# prepare some data for A and B groups
dt <- data.frame(
  by = c(rep("A", 15), rep("B",10))
)

adt <- MASS::mvrnorm(15, c(0,0), cbind(c(1,0.8), c(0.8,1)))
bdt <- MASS::mvrnorm(10, c(0,0), cbind(c(1,0.1), c(0.1,1)))

mat <- rbind(adt, bdt)
colnames(mat) <- c("x", "y")
dt <- cbind(dt, mat)

res <- continuousTestByVariable(dt, by = "by", x = "x", y = "y")
res <- res \%>\% arrange(by)

atest <- cor.test(adt[,1], adt[,2])
btest <- cor.test(bdt[,1], bdt[,2])
all.equal(res$pvalue, c(atest$p.value, btest$p.value))

}
