% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{createGCloudAccessToken}
\alias{createGCloudAccessToken}
\title{Create GCloud access token}
\usage{
createGCloudAccessToken()
}
\value{
list with fields: user, token, timestamp
}
\description{
Retrieves a GCloud access token for the currently logged-in user
}
