% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{createMachineLearningModel}
\alias{createMachineLearningModel}
\title{Create machine learning model using XIFF package.}
\usage{
createMachineLearningModel(
  trainingSet,
  geneSet,
  geneAnno,
  p = FALSE,
  classLabel = list(class1_name = "class1", class2_name = "class2"),
  method = "rf",
  tuneLength = 5,
  number = 10,
  repeats = 10,
  selectBestFeaturesFnc = selectBestFeatures,
  threshold = "Confirmed",
  maxFeatures = "auto",
  ...
)
}
\description{
Create machine learning model using XIFF package.
}
\examples{

data("data_createMachineLearningModel", package = "XIFF")
trainingSet <- data_createMachineLearningModel$trainingSet
geneSet     <- data_createMachineLearningModel$geneSet
geneAnno    <- data_createMachineLearningModel$geneAnno

fit <- createMachineLearningModel(trainingSet, geneSet, geneAnno)
fitNN <- createMachineLearningModel(trainingSet, geneSet, geneAnno, method = "neuralnetwork")

}
