% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{differentialBayesCommon}
\alias{differentialBayesCommon}
\title{Internal Function to be used differentialXXX family of functions}
\usage{
differentialBayesCommon(
  sampleClasses,
  dbDataFun,
  idCol,
  scoreCol,
  progressLabel,
  itemLabel,
  p = FALSE
)
}
\arguments{
\item{sampleClasses}{classAssignment object}

\item{dbDataFun}{function to fetch the data from database}

\item{idCol}{name of the column with id}

\item{scoreCol}{name of the column with score}

\item{progressLabel}{optional label for progress bar}

\item{itemLabel}{optional label for progress bar}

\item{p}{progress type, logical or shiny session object}
}
\description{
Internal Function to be used differentialXXX family of functions
}
