% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dimensionReduction.R
\name{dimensionReductionUI}
\alias{dimensionReductionUI}
\alias{dimensionReduction}
\title{Dimension Reduction Module}
\usage{
dimensionReductionUI(id, width = "100\%", height = "800px")

dimensionReduction(
  input,
  output,
  session,
  InputData,
  AnalysisParams,
  ClusterMethod,
  PlotParams,
  classLabel,
  funTSNE = calcTSNE,
  funPCA = calcPCA_expression,
  funUMAP = calcUMAP,
  funPHATE = calcPHATE
)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{InputData}{}

\item{AnalysisParams}{}

\item{ClusterMethod}{}

\item{PlotParams}{}

\item{classLabel}{}

\item{funTSNE}{}

\item{funPCA}{}

\item{funUMAP}{}

\item{funPHATE}{}
}
\description{
Dimension Reduction Module
}
