% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dimensionReduction.R
\name{dimRedAvailableMethods}
\alias{dimRedAvailableMethods}
\title{Returns the list of supported dimension reduction methods.}
\usage{
dimRedAvailableMethods()
}
\value{
character vector with supported methods.
}
\description{
Returns the list of supported dimension reduction methods.
}
