% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-params.R
\name{doubleClassTestByVariable}
\alias{doubleClassTestByVariable}
\title{Double Class Test}
\usage{
doubleClassTestByVariable(
  data,
  byVar,
  x,
  y,
  test = matrixTests::row_t_equalvar,
  ...
)
}
\arguments{
\item{data}{data.frame containing \code{byVar}, \code{x}, \code{y} columns.}

\item{byVar}{a column name used to split the data for tests.}

\item{x}{column name which contains two levels used to split \code{y} column.}

\item{y}{column name from \code{data}, which contains continuous variable.}

\item{test}{function used to perform the tests. It needs to take two matrices.}

\item{...}{other params passed to test function.}
}
\value{
data.frame with statistical tests applied across all the \code{byVar} levels.
}
\description{
Double Class Test
}
\examples{

set.seed(123)
dt <- data.frame(
  by = c(rep("A", 15), rep("B",10)),
  x  = c(
     rep("a",7), rep("b",8), # A
     rep("a",6), rep("b",4)  # B
   ),
  y  = c(
     rnorm(7), rnorm(8, 2),  # A
     rnorm(6), rnorm(4, 4)  # B
   )
)

res <- doubleClassTestByVariable(dt, byVar = "by", x = "x", y = "y")
res <- res \%>\% arrange(by)

atest <- t.test(y~x, data = dt \%>\% filter(by == "A"), var.equal = TRUE)
btest <- t.test(y~x, data = dt \%>\% filter(by == "B"), var.equal = TRUE)
all.equal(res$pvalue, c(atest$p.value, btest$p.value))

}
