% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dropNulls}
\alias{dropNulls}
\title{Drop NULL values from list}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{list}
}
\value{
list x without NULL values
}
\description{
Drop NULL values from list
}
\examples{

ls <- list(1, NULL, 2)
dropNulls(ls)

}
