% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{ensureCommonRownames}
\alias{ensureCommonRownames}
\title{Extract rows with the same rownames from data.frame}
\usage{
ensureCommonRownames(m1, m2, sortRownames = FALSE, outNames = NULL)
}
\arguments{
\item{m1}{first data frame}

\item{m2}{second data.frame}

\item{sortRownames}{should the rownames should sorted}

\item{outNames}{names for the output list}
}
\value{
list with m1 and m2 data.frame both containing rows with the same names
}
\description{
Extract rows with the same rownames from data.frame
}
\examples{

m1 <- data.frame(x = 1:3, y = 1:3, row.names = c("A", "B", "C"))
m2 <- data.frame(x = 2:4*100, y = 2:4*10, row.names = c("B", "C", "D"))

ensureCommonRownames(m1,m2)

}
