% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exampleClassAssigment}
\alias{exampleClassAssigment}
\title{Example Class Assignment to be used in Examples}
\usage{
exampleClassAssigment()
}
\value{
classAssignment object
}
\description{
Example Class Assignment to be used in Examples
}
\examples{

exampleClassAssigment()

}
