% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exampleDataFrame}
\alias{exampleDataFrame}
\title{Example Data Frame to be used in Examples}
\usage{
exampleDataFrame()
}
\value{
data.frame object
}
\description{
Example Data Frame to be used in Examples
}
\examples{

exampleDataFrame()

}
