% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_expression_volcano.R
\name{expressionVolcanoTabUI_main}
\alias{expressionVolcanoTabUI_main}
\alias{expressionVolcanoTabUI_sidebar}
\alias{expressionVolcanoTab}
\title{Expression Volcano Tab}
\usage{
expressionVolcanoTabUI_main(id)

expressionVolcanoTabUI_sidebar(id)

expressionVolcanoTab(input, output, session, classLabel, TableData)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{classLabel}{}

\item{TableData}{}
}
\description{
Expression Volcano Tab
}
