% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_extendedTableDownload.R
\name{extendedTableDownloadUI}
\alias{extendedTableDownloadUI}
\alias{extendedTableDownload}
\title{Module for downloading tables as csv files.}
\usage{
extendedTableDownloadUI(id, label, filename = label)

extendedTableDownload(
  input,
  output,
  session,
  Table,
  Subject,
  Item,
  classSelection,
  classLabel,
  by,
  additional = NULL
)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{Table}{}

\item{Subject}{}

\item{Item}{}

\item{classSelection}{}

\item{classLabel}{}

\item{by}{}

\item{additional}{}
}
\description{
Module for downloading tables as csv files.
}
