% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmButtonOneClass}
\alias{fmButtonOneClass}
\alias{fmButtonBothClasses}
\title{Utility function for styling Future Manager Run Buttons}
\usage{
fmButtonOneClass(inputId, fm, cs, defaultValue = FALSE)

fmButtonBothClasses(inputId, fm, cs, defaultValue = FALSE)
}
\arguments{
\item{inputId}{module id}

\item{fm}{FutureManager object}

\item{cs}{class selection object}

\item{defaultValue}{logical, the initial button value}
}
\description{
Utility function for styling Future Manager Run Buttons
}
