% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{gatherPredictionResults}
\alias{gatherPredictionResults}
\title{gatherPredictionResults}
\usage{
gatherPredictionResults(predictions, ...)
}
\description{
gatherPredictionResults
}
\details{
TODO: is this function even needed?
}
