% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_dimRed.R
\name{geneExpressionDimRed}
\alias{geneExpressionDimRed}
\title{Gene Expression Dimension Reduction}
\usage{
geneExpressionDimRed(
  geneExpr,
  anno,
  method = XIFF::dimRedAvailableMethods(),
  clusterMethod = c("pam k = automatic", "pam k = 2", "pam k = 3", "pam k = 4",
    "affinity propagation", "k-means k = automatic", "k-means k = 2", "k-means k = 3",
    "k-means k = 4"),
  useCorrelation = TRUE,
  ...,
  ca = NULL
)
}
\arguments{
\item{geneExpr}{gene expression data}

\item{anno}{gene annotation}

\item{method}{dimension reduction method}

\item{clusterMethod}{clustering method}

\item{useCorrelation}{should the dimension reduction be done on the corr matrix}

\item{...}{parameters passed to dimension reduction functions. See details.}

\item{ca}{}
}
\description{
Gene Expression Dimension Reduction
}
\details{
To check which additional arguments can be passed to specific dimension
reduction methods, please refer to their documentation:

\enumerate{
\item tsne - \link[Rtsne:Rtsne]{Rtsne::Rtsne}
\item pca - \link{prcomp}
\item umap - \link[umap:umap]{umap::umap}
\item phate - \link[phateR:phate]{phateR::phate}
}
}
