% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_dimRed.R
\name{geneExpressionDimRedPlot}
\alias{geneExpressionDimRedPlot}
\title{Create Gene Expression Dimension Reduction Plot}
\usage{
geneExpressionDimRedPlot(dimRed, useLabels = FALSE)
}
\arguments{
\item{dimRed}{result of the \code{geneExpressionDimRed} function}

\item{useLabels}{logical, if true add labels to the plot}
}
\description{
Create Gene Expression Dimension Reduction Plot
}
