% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_geneExpression_dimRed.R
\name{geneExpressionDimRedTabUI_main}
\alias{geneExpressionDimRedTabUI_main}
\alias{geneExpressionDimRedTabUI_sidebar}
\alias{geneExpressionDimRedTab}
\title{Gene Expression Dimension Reduction Module}
\usage{
geneExpressionDimRedTabUI_main(id)

geneExpressionDimRedTabUI_sidebar(id)

geneExpressionDimRedTab(
  input,
  output,
  session,
  fm,
  Results,
  TableData,
  getDataGeneExpression
)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{fm}{}

\item{Results}{}

\item{TableData}{}

\item{getDataGeneExpression}{}
}
\description{
Gene Expression Dimension Reduction Module
}
