% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_plots.R
\name{generateApplyPerformancePlot}
\alias{generateApplyPerformancePlot}
\title{Generate Apply Performance Plot}
\usage{
generateApplyPerformancePlot(df)
}
\arguments{
\item{df}{MLModelTestsResult object or 
the result of \code{getPerformanceDataFrame}}
}
\value{
ggplot object
}
\description{
Generate Apply Performance Plot
}
\examples{

confMatrix <- structure(
  c(19L, 9L, 2L, 66L),
  .Dim = c(2L, 2L),
  .Dimnames = list(
    Prediction = c("positive", "negative"),
    Reference = c("positive",
                  "negative")
  ),
  class = "table"
)
df <- getPerformanceDataFrame(confMatrix)
generateApplyPerformancePlot(df)

}
