% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateClassSelectionPlot}
\alias{generateClassSelectionPlot}
\title{Generate Class Selection Plot}
\usage{
generateClassSelectionPlot(
  sampleClasses,
  prop1,
  prop2,
  n_classes,
  plot_type = "bar",
  usePercent = FALSE,
  annotation,
  annotationFocus,
  classLabel = NULL
)
}
\arguments{
\item{sampleClasses}{classAssigment object}

\item{prop1}{property to be put on x axis}

\item{prop2}{property used for color scale}

\item{n_classes}{}

\item{plot_type}{"bar" or "pie"}

\item{usePercent}{if TRUE uses percentages rather than counts}

\item{annotation}{annotaion data}

\item{annotationFocus}{annotation data}

\item{classLabel}{optional list with class labels}
}
\description{
Generate Class Selection Plot
}
\examples{

if(require("CLIFF")) {
  CLIFF::setDbOptions()
  cs <- CLIFF::exampleClassAssigment()
  anno <- CLIFF::getCellLineAnno("human")
  
  generateClassSelectionPlot(
    sampleClasses = cs,
    prop1 = "tumortype", prop2 = "gender", n_classes = 10,
    annotation = anno, annotationFocus = anno
  )
  
  generateClassSelectionPlot(
    sampleClasses = cs,
    prop1 = "tumortype", prop2 = "gender", n_classes = 10,
    annotation = anno, annotationFocus = anno,
    usePercent = TRUE
  )
}

}
