% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_plots.R
\name{generateErrorPlot}
\alias{generateErrorPlot}
\title{Generate Error Plot for XIFF Random Forrest Machine Learning Model}
\usage{
generateErrorPlot(
  x,
  cl = list(class1_name = "sensitive", class2_name = "resistant")
)
}
\arguments{
\item{x}{Xiff Machine Learning MOdel}

\item{cl}{list with class labels}
}
\value{
ggplot2 object containing error plot
}
\description{
Generate Error Plot for XIFF Random Forrest Machine Learning Model
}
