% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_plots.R
\name{generatePerformancePlot}
\alias{generatePerformancePlot}
\title{Generate Machine Learning Performance Plot}
\usage{
generatePerformancePlot(x)
}
\description{
Generate Machine Learning Performance Plot
}
\examples{
\dontrun{
# TODO: create models here
generatePerformancePlot(fit$trainingOutput)
generatePerformancePlot(fitNN$trainingOutput)

}

}
