% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateScoreBarPlot}
\alias{generateScoreBarPlot}
\title{Generate Score Bar Plot}
\usage{
generateScoreBarPlot(data, score_desc)
}
\arguments{
\item{data}{data.frame with \code{x_score} column}

\item{score_desc}{}
}
\value{
ggplot2 plot
}
\description{
Generate Score Bar Plot
}
\examples{

dt <- exampleDataFrame()
dt <- dt \%>\% mutate(x_score = cut(tpm, breaks = c(0,5,10,20,50, Inf)))

generateScoreBarPlot(dt, "TMP")

}
