% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateScoreWaterfallPlot}
\alias{generateScoreWaterfallPlot}
\title{Generate Score Waterfall plot}
\usage{
generateScoreWaterfallPlot(data, score_desc, y_scale = "identity")
}
\arguments{
\item{data}{data.frame containing \code{x_score} and \code{tumortype} column}

\item{score_desc}{}

\item{y_scale}{type of y-scale (e.g. identity or log10)}
}
\value{
ggplot2 plot
}
\description{
Generate Score Waterfall plot
}
\examples{

df <- exampleDataFrame() \%>\% reorderByScore(valueCol = "tpm")
df <- df \%>\% rename(x_score = tpm) \%>\%
  mutate(tumortype = c(rep("A", 5), rep("B", 6)))
generateScoreWaterfallPlot(df, "TPM")
generateScoreWaterfallPlot(df, "TPM", "log10")

}
