% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_plots.R
\name{generateTablePlot}
\alias{generateTablePlot}
\title{Generate Visualization of the confusion matrix.}
\usage{
generateTablePlot(df)
}
\arguments{
\item{df}{MLModelTestsResult or data.frame with columns predicted and reference columns}
}
\value{
ggplot object
}
\description{
Generate Visualization of the confusion matrix.
}
\examples{

lvls <- factor(
  c("resistant\n(negative)", "sensitive\n(positive)"),
  ordered = TRUE
)

tbl <- tibble(predicted = lvls[c(1,1,1,2,2)], reference = lvls[c(2,1,1,1,2)])

generateTablePlot(tbl)

}
