% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{generateTestModelPlots}
\alias{generateTestModelPlots}
\title{Generate Machine Learning Model based on the Test Results.}
\usage{
generateTestModelPlots(testResult)
}
\arguments{
\item{testResults}{MLModelTestsResult object created with testModel function.}
}
\value{
list with plots created on test result.
}
\description{
Generate Machine Learning Model based on the Test Results.
}
