% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{generateTestPerformanceData}
\alias{generateTestPerformanceData}
\title{Get Performance Data Frame}
\usage{
generateTestPerformanceData(result)
}
\arguments{
\item{result}{MLModelTestsResult object or 
confusion matrix (table object) or}
}
\value{
table with columns: metric, n1, n2, value, labelPos, label
}
\description{
Get Performance Data Frame
}
\examples{

confMatrix <- as.table(matrix(
  c(19L, 9L, 2L, 66L),
  nrow = 2,
  dimnames = list(
    Prediction = c("positive", "negative"),
    Reference =  c("positive", "negative")
  )
))
generateTestPerformanceData(confMatrix)

}
