% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{generateTrainingModelPlots}
\alias{generateTrainingModelPlots}
\title{Generate Machine Learning Models Plots on Training Set}
\usage{
generateTrainingModelPlots(model)
}
\arguments{
\item{model}{machine learning model created with XIFF (MLXIFF object)}
}
\value{
list with plots created on based on the training set.
}
\description{
Generate Machine Learning Models Plots on Training Set
}
