% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_plots.R
\name{generateVarImpPlot}
\alias{generateVarImpPlot}
\title{Create Machine Learning Variable Importance Plot.}
\usage{
generateVarImpPlot(x)
}
\description{
Create Machine Learning Variable Importance Plot.
}
\examples{

\dontrun{
# TODO: create models here
generateVarImpPlot(fit)
generateVarImpPlot(fitNN)

}

}
