% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateVolcanoPlot}
\alias{generateVolcanoPlot}
\title{Generate Volcano Plot}
\usage{
generateVolcanoPlot(
  diffExpr,
  minuslog10pval,
  minuslog10adjpval,
  log2FC,
  FCside = "both sides",
  classLabels = c("class1", "class2")
)
}
\arguments{
\item{diffExpr}{differential result expression table}

\item{minuslog10pval}{}

\item{minuslog10adjpval}{}

\item{log2FC}{}

\item{FCside}{"both sides", "left", "right"}

\item{classLabels}{}
}
\value{
ggplot2 object
}
\description{
Generate Volcano Plot
}
\examples{

if(require("CLIFF")) {
   gene_anno <- CLIFF::getGeneAnno("human")
   cs <- CLIFF::exampleClassAssigment()
   limmaData <- CLIFF::differentialGeneExpression_LimmaVoom(cs, gene_anno)
   generateVolcanoPlot(
     limmaData,
     minuslog10pval = 2,
     minuslog10adjpval = 0,
     log2FC = 2.5,
     FCside = "both sides"
   )
}


}
