% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateWaterfallPlot}
\alias{generateWaterfallPlot}
\title{Generate Waterfall Plot}
\usage{
generateWaterfallPlot(
  data,
  dataCol,
  xlabel = getOption("xiff.label"),
  ylabel = "score",
  trans = "identity",
  limits = NULL,
  fill = "tumortype"
)
}
\arguments{
\item{data}{}

\item{dataCol}{}

\item{xlabel}{}

\item{ylabel}{}

\item{trans}{}

\item{limits}{}

\item{fill}{}
}
\value{

}
\description{
Generate Waterfall Plot
}
\examples{

df <- exampleDataFrame() \%>\% reorderByScore(valueCol = "tpm")
generateWaterfallPlot(df, "tpm", fill = "tpm")

}
