% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getEnsemblLocationLink}
\alias{getEnsemblLocationLink}
\alias{getEnsemblGeneLink}
\alias{getEnsemblTranscriptLink}
\alias{getGeneSetLink}
\title{Get HTML tags with links}
\usage{
getEnsemblLocationLink(location, species = "human")

getEnsemblGeneLink(gene, species = "human")

getEnsemblTranscriptLink(transcript, species)

getGeneSetLink(geneset, msigDBLink, rawGeneset = NULL)
}
\arguments{
\item{location}{}

\item{species}{}
}
\value{

}
\description{
Get HTML tags with links
}
