% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_label}
\alias{get_label}
\title{Get XIFF label with the number of items for plots}
\usage{
get_label(n)
}
\arguments{
\item{n}{number of items}
}
\value{
string with the label
}
\description{
Get XIFF label with the number of items for plots
}
\examples{

get_label(50)

}
