% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_bioinfo.R
\name{getAntibodyInformation}
\alias{getAntibodyInformation}
\title{Get Antibody Information}
\usage{
getAntibodyInformation(antibody = NA, checkData = FALSE)
}
\arguments{
\item{antibody}{single character string for which to return the information
or \code{NA} (default), to return data for all antibodies.}

\item{checkData}{if true check for data avability in 
\code{processedproteinexpression} table and adds \code{has_data} column
with logical indicator}
}
\value{
data.frame with antibody, validation_status, vendor, catalog_number
columns and optionally if checkData is TRUE, logical column has_data which 
determines if specific antibody has data available in table
\code{processedproteinexpression}; 
if parameter \code{antibody} is a single value the result will also contain
column \code{genes}
}
\description{
Get Antibody Information
}
