% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAssignmentDf}
\alias{getAssignmentDf}
\title{Convert ClassAssigment to data.frame}
\usage{
getAssignmentDf(ca, useLabels = TRUE, returnFactor = TRUE)
}
\arguments{
\item{ca}{classAssigment}

\item{useLabels}{use classAssigment labels}

\item{returnFactor}{logical. If true (default), then class column will be
returned as factor}
}
\value{
data.frame with two columns
}
\description{
Convert ClassAssigment to data.frame
}
\examples{

ca <- exampleClassAssigment()
getAssignmentDf(ca)

setClassLabel(ca, "class2") <- "C2"
getAssignmentDf(ca)

}
