% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{dropUnbalancedVariableValues}
\alias{dropUnbalancedVariableValues}
\alias{dropUnbalancedTumortypes}
\alias{getValidTumorTypes}
\alias{getBalancedTumorTypes}
\alias{getBalancedTumortypes}
\alias{getBalancedVariableValues}
\title{Get valid tumor types / variables}
\usage{
dropUnbalancedVariableValues(cs, anno, variable, minCount = 1)

dropUnbalancedTumortypes(cs, anno, minCount = 1)

getValidTumorTypes(cs, anno, minCount = 1)

getBalancedTumorTypes(cs, anno, minCount = 1)

getBalancedTumortypes(cs, anno, minCount = 1)

getBalancedVariableValues(cs, anno, variable, minCount = 1)
}
\arguments{
\item{cs}{list, class selection}

\item{anno}{data.frame, item annotation}

\item{variable}{column name from which the balanced values will be extracted}

\item{minCount}{minimum number of occurrences in each class to consider a level
as valid. If \code{minCount} is lower that 1, then all levels will
be returned (except NAs).}
}
\value{
character vector of valid tumortypes
}
\description{
This function returns tumor types that are available in the data for both classes
}
\examples{
anno <- tibble(
 cl = c("a", "b", "c", "d", "e", "f", "g"),
 tumortype = c("x", "x", "y", "x", "x", "y", "z")
)
colnames(anno)[1] <- getOption("xiff.column")

cs <- classAssignment(cl1 = c("a", "b", "c"), cl2 = c("d", "e", "f", "g"))

getBalancedTumortypes(cs, anno) # "x" and "y" are available in both classes

# 'g' has unbalanced tumortype
dropUnbalancedTumortypes(cs, anno)
dropUnbalancedTumortypes(cs, anno)$class2

# only "x" appears at least 2 times in each class
getBalancedTumortypes(cs, anno, minCount = 2) 

cs2 <- dropUnbalancedTumortypes(cs, anno, 2)
cs2$class1
cs2$class2

# empty results
getBalancedTumortypes(cs, anno, 3)
dropUnbalancedTumortypes(cs, anno, 3)

# getBalancedVariableValues and dropUnbalancedVariableValues
getBalancedVariableValues(cs, anno, "tumortype", 1)
getBalancedVariableValues(cs, anno, "tumortype", 2)
dropUnbalancedVariableValues(cs, anno, "tumortype", 2)

# minCount == 0
getBalancedVariableValues(cs, anno, "tumortype", 0)
dropUnbalancedVariableValues(cs, anno, "tumortype", 0)

}
