% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{getClassDistances}
\alias{getClassDistances}
\title{Get Class Distances}
\usage{
getClassDistances(mat, pheno, metric = "euclidean")
}
\arguments{
\item{mat}{gene set matrix (e.g. result of \code{getHeatmapDataHallmark})}

\item{pheno}{data.frame with class assigment}

\item{metric}{metrix used to calculate distances}
}
\value{
distance matrix
}
\description{
Get Class Distances
}
\examples{

if(require("CLIFF")) {
  
  CLIFF::setDbOptions()
  cs <- CLIFF::exampleClassAssigment()
  data <- CLIFF::getHeatmapDataHallmark(cs)
  pheno <- stackClasses(cs)
  
  res <- getClassDistances(data, pheno)
  
  as.matrix(res)[1:3, 1:3]

}

}
