% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getClassLabel}
\alias{getClassLabel}
\title{Get class label list from classAssignment object.}
\usage{
getClassLabel(ca)
}
\arguments{
\item{ca}{classAssignment object}
}
\description{
Get class label list from classAssignment object.
}
