% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{getCurrentGCloudUser}
\alias{getCurrentGCloudUser}
\title{Get current GCloud user}
\usage{
getCurrentGCloudUser()
}
\value{
the current user email
}
\description{
Returns a currently logged-in GCloud user
}
