% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{getDataForModel}
\alias{getDataForModel}
\title{Get Data for Machine Learning Model}
\usage{
getDataForModel(
  assignment,
  features,
  schema = getOption("xiff.schema"),
  column = getOption("xiff.column"),
  classLabel = NULL
)
}
\arguments{
\item{assignment}{stackClass assignment of items or classAssignment object}

\item{features}{character vector of genes ensg or fitted XIFF Machine Learning model}

\item{schema}{data base schema, default value should be used}

\item{column}{column containing the items, default value shoud be used}

\item{classLabel}{list with alternative class labels}
}
\value{
raw data.frame with features for MLXIFF model
}
\description{
Get Data for Machine Learning Model
}
\examples{
if(require("CLIFF")){
  CLIFF::setDbOptions()
  ca <- CLIFF::exampleClassAssigment()
  geneSet <- head(getGSEAdata("human", gene_set = "HALLMARK_P53_PATHWAY"), 3)

  modelData <- getDataForModel(ca, geneSet)
  head(modelData)

  getDataForModel(ca, geneSet, classLabel = list(class1 = "c1", class2 = "c2"))
}

}
