% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getExpressionDimRed}
\alias{getExpressionDimRed}
\title{Create Expression DimReduction.}
\usage{
getExpressionDimRed(
  data,
  anno,
  method = XIFF::dimRedAvailableMethods(),
  clusterMethod = c("pam k = automatic", "pam k = 2", "pam k = 3", "pam k = 4",
    "affinity propagation", "k-means k = automatic", "k-means k = 2", "k-means k = 3",
    "k-means k = 4"),
  geneSource = c("varying_genes", "gene_set"),
  numGenes = 30,
  unit = "log2tpm",
  ...,
  .p = TRUE
)
}
\arguments{
\item{data}{result of getExpressionDimRedData function.}

\item{anno}{data annotation.}

\item{method}{method for dimension reduction. 
See: \code{dimRedAvailableMethods()} for supported methods.}

\item{clusterMethod}{clustering method.}

\item{geneSource}{}

\item{numGenes}{}

\item{unit}{}

\item{...}{other parameters passed to dimension reduction functions.}

\item{.p}{}

\item{ca}{}
}
\description{
Create Expression DimReduction.
}
