% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{getFmStrategy}
\alias{getFmStrategy}
\title{Get future strategy}
\usage{
getFmStrategy(settings = NULL)
}
\arguments{
\item{settings}{application settings list}
}
\value{
future strategy
}
\description{
Returns strategy for the background calculations. Settings should contain
settings$options$fm_strategy field with a character string: "auto", "multicore",
or "multisession". Auto strategy will use 
\code{\link[FutureManager]{fmParallelStrategy}}, other choices refer to 
\code{\link[future]{multicore}} and \code{\link[future]{multisession}} respectively.
}
\details{
For debug purposes \code{\link[future]{sequential}} is used - to enable it please 
set `XIFF_FM_DEBUG` environmental variable to "YES".
}
