% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{getGCloudAccessToken}
\alias{getGCloudAccessToken}
\title{Get GCloud access token}
\usage{
getGCloudAccessToken()
}
\value{
list with fields: user, token, timestamp
}
\description{
Returns an active GCloud access token. If the token does not exist or is
outdated (was created more than 59min ago), a new one is created and stored
in xiff.gToken option.
}
