% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{getGenesSymbols}
\alias{getGenesSymbols}
\title{Get gene symbols}
\usage{
getGenesSymbols(ensgs, species = "human")
}
\arguments{
\item{ensgs}{gene ensg}

\item{species}{single character string determining the species}
}
\value{
vector of gene symbols. If there's no ensg in the database the NA is
returned for that gene.
}
\description{
Get gene symbols
}
\examples{

\dontrun{

getGenesSymbols(c("ENSG00000133703", "xx", "ENSG00000268173", "ENSG00000133703"))

}

}
