% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_affinity_ratio.R
\name{getGrepFeatureSelection}
\alias{getGrepFeatureSelection}
\title{Feature selection using GREP methodology.}
\usage{
getGrepFeatureSelection(
  df,
  threshold = 0.05,
  maxFeatures = 600,
  maxGenes = 300,
  minFeatures = 15,
  epsilonRNAseq = 10,
  cor.method = "pearson",
  .otherParams = list()
)
}
\arguments{
\item{df}{data frame with ensg.}

\item{threshold}{not used, can be anything. Rquired only for compability
with other feature selection methods.}

\item{maxFeatures}{max number of genes to be used in ratio calculations.}

\item{minFeatures}{min number of features resulting from affinity propagation.}

\item{epsilonRNAseq}{}
}
\description{
Feature selection using GREP methodology.
}
