% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_bioinfo.R
\name{getGSEAdata}
\alias{getGSEAdata}
\title{Get gene set data.}
\usage{
getGSEAdata(species, collection_name = "mSigDB", gene_set = NULL)
}
\arguments{
\item{species}{selected species. E.g. 'human'.}

\item{collection_name}{name of the gene set collection.}

\item{gene_set}{name of single gene set to be returned. If specified then the
function returns character vector containing that gene set.}
}
\value{
List of gene set. If \code{gene_set} is used, then it returns just a 
character vector with that gene set.
}
\description{
Get gene set data.
}
