% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getHallmarkGeneSetChoices}
\alias{getHallmarkGeneSetChoices}
\title{Get Hallmark Gene Set Choices}
\usage{
getHallmarkGeneSetChoices(geneSets)
}
\arguments{
\item{geneSets}{vector with HALLMARK genes setes names}
}
\value{

}
\description{
Get Hallmark Gene Set Choices
}
\examples{

if(require("CLIFF")) {

   dt <- CLIFF::differentialHallmarkSets(CLIFF::exampleClassAssigment()) \%>\%
     head \%>\% select(gene_set)
   names(getHallmarkGeneSetChoices(dt$gene_set))

}

}
