% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{getPerformanceDataFrame}
\alias{getPerformanceDataFrame}
\title{Get Performance Data Frame}
\usage{
getPerformanceDataFrame(result)
}
\arguments{
\item{result}{MLModelTestsResult object or 
confusion matrix (table object) or}
}
\value{
table with columns: metric, n1, n2, value, labelPos, label
}
\description{
Get Performance Data Frame
}
\examples{

resTable <- structure(
  c(19L, 9L, 2L, 66L),
  .Dim = c(2L, 2L),
  .Dimnames = list(
    Prediction = c("positive", "negative"),
    Reference = c("positive",
                  "negative")
  ),
  class = "table"
)
getPerformanceDataFrame(resTable)

}
