% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{getPostgresql}
\alias{getPostgresql}
\title{Get data from Postgres}
\usage{
getPostgresql(sql)
}
\arguments{
\item{sql}{character string, the SQL query}
}
\value{
data.frame
}
\description{
Function that queries the DB
}
