% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{getPostgresqlConnection}
\alias{getPostgresqlConnection}
\title{Get PostgreSQL Connection object}
\usage{
getPostgresqlConnection()
}
\value{
RPostgres connection object
}
\description{
This function allows to get a DB connection object, that can be used in dbplyr
pipelines.
}
\examples{

setDbOptions()
con <- getPostgresqlConnection()
con \%>\%
  dplyr::tbl("msigdb") \%>\%
  head() \%>\%
  dplyr::collect()
RPostgres::dbDisconnect(con)

}
