% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{getRawDataForModel}
\alias{getRawDataForModel}
\title{Get Raw Data For MLXIFF Model}
\usage{
getRawDataForModel(
  features,
  names = NULL,
  schema = getOption("xiff.schema"),
  column = getOption("xiff.column")
)
}
\arguments{
\item{features}{character vector with features or MLXIFF object}

\item{names}{character vector with items for which the data needs to be extracted}

\item{schema}{data base schema to be used
(in most cases default value is sufficient)}

\item{column}{name of the column which stores the items
(in most cases default value is sufficient)}
}
\value{
long form data.frame containing {itemname}, ensg, and score
}
\description{
Get Raw Data For MLXIFF Model
}
\examples{

if(require("CLIFF")) {

  ca <- CLIFF::exampleClassAssigment()
  geneSet <- getGSEAdata("human", "hallmark", "HALLMARK_P53_PATHWAY")
  getRawDataForModel(geneSet, unlist(ca))

}

}
