% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{getValidTumorTypes}
\alias{getValidTumorTypes}
\title{Get valid tumor types}
\usage{
getValidTumorTypes(cs, anno)
}
\arguments{
\item{cs}{list, class selection}

\item{anno}{data.frame, item annotation}
}
\value{
character vector of valid tumortypes
}
\description{
This function returns tumor types that are available in the data for both classes
}
