% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{ggColorHue}
\alias{ggColorHue}
\title{XIFF Color Scale}
\usage{
ggColorHue(n)
}
\arguments{
\item{n}{number of colors to be returned}
}
\description{
XIFF Color Scale
}
\examples{

cols <- ggColorHue(6)
pie(rep(1,6), col = cols)
pie(rep(1,200), col = ggColorHue(200), lty = 0, labels = "")


}
