% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{guessAxisScale}
\alias{guessAxisScale}
\title{Determine appropriate scale type for given set of values}
\usage{
guessAxisScale(values)
}
\arguments{
\item{values}{vector with values}
}
\value{
preferred scale to be used for given values
}
\description{
Determine appropriate scale type for given set of values
}
\examples{

x <- data.frame(x = 1:3, v = c(1,2,30))
y <- data.frame(x = 1:4, v = c(1,2,30,200))


guessAxisScale()
guessAxisScale()

library(ggplot2)

#identity
ggplot(x) + geom_bar(aes(x = x, y = v), stat = "identity") + 
 scale_y_continuous(trans = guessAxisScale(x$v))
 
# log10
ggplot(y) + geom_bar(aes(x = x, y = v), stat = "identity") + 
 scale_y_continuous(trans = guessAxisScale(y$v))


}
