% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{initialDbCheck}
\alias{initialDbCheck}
\title{Check Database Connection}
\usage{
initialDbCheck(timeout = 5)
}
\arguments{
\item{timeout}{timout for database connection}
}
\value{
logical, TRUE if database connection works, or FALSE otherwise
}
\description{
Check Database Connection
}
