% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{inputWithHelp}
\alias{inputWithHelp}
\alias{helpIcon}
\title{Add help text to selectInput shiny object}
\usage{
inputWithHelp(param, helpText = "help me Obi-Wan Kenobi")

helpIcon(text)
}
\arguments{
\item{param}{selectInput shiny object}

\item{helpText}{}
}
\description{
Add help text to selectInput shiny object
}
\examples{

select <- shiny::selectInput("x", "Input", 1:2)
select2 <- inputWithHelp(select, "Help!")

if(interactive()) {
  shinyApp(fluidPage(select2), function(input, output, session) {})
}

}
