% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{invalid}
\alias{invalid}
\title{Test if value is empty string, NA or NULL}
\usage{
invalid(x)
}
\arguments{
\item{x}{any R object}
}
\value{
logical value
}
\description{
Test if value is empty string, NA or NULL
}
\examples{

invalid("")
invalid(NULL)
invalid(NA)

}
