% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{isDbOnline}
\alias{isDbOnline}
\title{Check if database is online}
\usage{
isDbOnline(timeout = 5)
}
\arguments{
\item{timeout}{database timeout in seconds}
}
\value{
logical, TRUE if database can be reached
}
\description{
Check if database is online
}
\examples{

isDbOnline()

}
