% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isDifferent}
\alias{isDifferent}
\title{Check if two object are different}
\usage{
isDifferent(x, y)
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\value{
true if x and y are different
}
\description{
Check if two object are different
}
\examples{

isDifferent(1, NULL)
isDifferent(1, 1)

}
