% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{isGCloudAvailable}
\alias{isGCloudAvailable}
\title{Check if Google Cloud SDK is installed}
\usage{
isGCloudAvailable()
}
\value{
logical, TRUE if Google Cloud if available in the search path
}
\description{
Check if Google Cloud SDK is installed
}
\examples{

isGCloudAvailable()

}
