% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_machineLearning_create.R
\name{machineLearningCreateModelTabUI_sidebar}
\alias{machineLearningCreateModelTabUI_sidebar}
\alias{machineLearningCreateModelTab}
\title{Title}
\usage{
machineLearningCreateModelTabUI_sidebar(id)

machineLearningCreateModelTab(
  input,
  output,
  session,
  fm,
  classSelection,
  classLabel,
  gsea_data_hallmark,
  gene_anno,
  AnnotationFocus,
  Species
)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{fm}{}

\item{classSelection}{}

\item{classLabel}{}

\item{gsea_data_hallmark}{}

\item{gene_anno}{}

\item{AnnotationFocus}{}

\item{Species}{}
}
\description{
Title
}
\examples{

}
