% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_machineLearning.R
\name{machineLearningTabUI_main}
\alias{machineLearningTabUI_main}
\alias{machineLearningTabUI_sidebar}
\alias{machineLearningTab}
\title{Title}
\usage{
machineLearningTabUI_main(id)

machineLearningTabUI_sidebar(id)

machineLearningTab(
  input,
  output,
  session,
  fm,
  classSelection,
  classLabel,
  gsea_data_hallmark,
  gene_anno,
  AnnotationFocus,
  Species
)
}
\arguments{
\item{id}{}

\item{input}{}

\item{output}{}

\item{session}{}

\item{fm}{}

\item{classSelection}{}

\item{classLabel}{}

\item{gsea_data_hallmark}{}

\item{gene_anno}{}

\item{AnnotationFocus}{}

\item{Species}{}
}
\description{
Title

Title

Title
}
