% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeClassAssignment}
\alias{makeClassAssignment}
\title{Utility function for creating class assignment objects.}
\usage{
makeClassAssignment(sampleClasses, classLabel)
}
\arguments{
\item{sampleClasses}{}

\item{classLabel}{}
}
\value{

}
\description{
Utility function for creating class assignment objects.
}
