% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{makeMlModelPlots}
\alias{makeMlModelPlots}
\title{Create all ML plots in one go.}
\usage{
makeMlModelPlots(model, testSet, annoFocus)
}
\arguments{
\item{model}{MLXIFF model}

\item{testSet}{test set}

\item{annoFocus}{annoFocus}
}
\description{
Create all ML plots in one go.
}
