% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{mlGetClassColumn}
\alias{mlGetClassColumn}
\title{Get class column name from MLXIFF model.}
\usage{
mlGetClassColumn(model, data = NULL, asSymbol = FALSE)
}
\arguments{
\item{model}{MLXIFF object.}

\item{asSymbol}{logical. If FALSE returns string. If true returns symbol.}
}
\value{

}
\description{
Get class column name from MLXIFF model.
}
