% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{mlGetTableData}
\alias{mlGetTableData}
\title{Summarise Model Parameters in Table}
\usage{
mlGetTableData(model)
}
\arguments{
\item{model}{machine learning model fitted in XIFF}
}
\value{

}
\description{
Summarise Model Parameters in Table
}
