% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{mlGetTpmData}
\alias{mlGetTpmData}
\title{Get TPM data for given ML Model}
\usage{
mlGetTpmData(model, ensg, annoFocus)
}
\arguments{
\item{model}{MLXIFF object}

\item{ensg}{gene ensg}

\item{annoFocus}{celline annotation data}
}
\value{
data.frame with columns: celllinename, ensg, tpm, tumortype
}
\description{
Get TPM data for given ML Model
}
\details{
This function is required to be implemented for new MLXIFF model.
Seel \code{mlGetTableData.XiffGREP} for an example.
}
