% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{mlSets2OriginalNames}
\alias{mlSets2OriginalNames}
\title{Rename column name and levels in the ml set to original values.}
\usage{
mlSets2OriginalNames(set, classColumn, classLabel)
}
\description{
Rename column name and levels in the ml set to original values.
}
