% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n_common}
\alias{n_common}
\title{Number of common entries in two vectors}
\usage{
n_common(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
interger denoting the number of common items in both vectors
}
\description{
Number of common entries in two vectors
}
\examples{

n_common(1:4,2:5)
n_common(1:4,5:10)

}
