% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Null default}
\usage{
x \%||\% y
}
\arguments{
\item{x}{any R object}

\item{y}{defalut value used if x is NULL}
}
\description{
Null default
}
\examples{

1 \%||\% 2
NULL \%||\% 2

}
