% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{packageInstalled}
\alias{packageInstalled}
\title{Check if Package is Available}
\usage{
packageInstalled(name)
}
\arguments{
\item{name}{package name}
}
\value{
logical, \code{TRUE} if package is available
}
\description{
Check if Package is Available
}
\examples{

packageInstalled("XIFF")
packageInstalled("XIFFFF")

}
