% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{plotColors}
\alias{plotColors}
\title{XIFF default color scheme}
\format{
An object of class \code{character} of length 7.
}
\usage{
plotColors
}
\description{
XIFF default color scheme
}
\keyword{datasets}
