% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plotWrapper.R
\name{plotWrapperUI}
\alias{plotWrapperUI}
\alias{plotWrapper}
\title{Plot Wrapper}
\usage{
plotWrapperUI(id, width = "100\%", height = "400px", center = FALSE)

plotWrapper(
  input,
  output,
  session,
  PlotExpr,
  PlotType = FALSE,
  filename = "image",
  varDict = list(class = "Class"),
  allowedTooltipTypes = c("point", "comparison", "scatterplot"),
  tooltipCallback = getOption("xiff.tooltipCallbackFun"),
  ...
)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{PlotExpr}{expression to create the plot}

\item{PlotType}{}

\item{filename}{}

\item{varDict}{}

\item{allowedTooltipTypes}{}

\item{tooltipCallback}{}

\item{...}{}
}
\description{
Plot Wrapper
}
