% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{prepareClassificationTable}
\alias{prepareClassificationTable}
\title{Prepare Classification Table}
\usage{
prepareClassificationTable(class1, class2, addRownames = TRUE)
}
\arguments{
\item{class1}{}

\item{class2}{}

\item{addRownames}{}
}
\value{

}
\description{
Prepare Classification Table
}
