% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{prepareConditionSql}
\alias{prepareConditionSql}
\title{Prepare SQL Where condition statements from list of R character vectors}
\usage{
prepareConditionSql(...)
}
\arguments{
\item{...}{character variables to be transformed into SQL WHERE statements}
}
\value{
string with SQL expression
}
\description{
Prepare SQL Where condition statements from list of R character vectors
}
\examples{

prepareConditionSql(x = c("1", "2"), y = c("A", "B"))

}
