% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{prepareDataAndDesignCommon}
\alias{prepareDataAndDesignCommon}
\title{Prepare Data And Design}
\usage{
prepareDataAndDesignCommon(sampleClasses, dbDataFun, idCol, scoreCol = "score")
}
\arguments{
\item{sampleClasses}{}

\item{dbDataFun}{}

\item{idCol}{}

\item{scoreCol}{}
}
\value{

}
\description{
Prepare Data And Design
}
