% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print.customPlotPrint}
\alias{print.customPlotPrint}
\title{Utility function for plotting customPlotPrint object}
\usage{
\method{print}{customPlotPrint}(x)
}
\arguments{
\item{x}{custom plot}
}
\value{

}
\description{
Utility function for plotting customPlotPrint object
}
