% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{print.MLModelTestsResult}
\alias{print.MLModelTestsResult}
\title{Print method for MLModelTestsResults}
\usage{
\method{print}{MLModelTestsResult}(x, ...)
}
\arguments{
\item{x}{MLModelTestsResults object}

\item{...}{discarded.}
}
\value{
invisible x
}
\description{
Print method for MLModelTestsResults
}
