% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{processMutationInfo}
\alias{processMutationInfo}
\title{Process Mutation Information}
\usage{
processMutationInfo(x, width)
}
\arguments{
\item{x}{a character vector or string to be processed.}

\item{width}{an integer specifying the desired maximum width of the text.}
}
\value{
a character vector with processed text.
}
\description{
This function processes input text to ensure that it doesn't exceed a specified width
and modify the semicolon in the text.
}
\examples{
x <- "c.100T>C;c.787G>T;c.840C>T;"
width <- 20
processed_text <- processMutationInfo(x, width)

}
