% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_progress.R
\name{ProcessProgress}
\alias{ProcessProgress}
\title{R6 ProcessProgress bar}
\description{
R6 ProcessProgress bar

R6 ProcessProgress bar
}
\details{
Progress bar for shiny
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProcessProgress-new}{\code{ProcessProgress$new()}}
\item \href{#method-ProcessProgress-update}{\code{ProcessProgress$update()}}
\item \href{#method-ProcessProgress-error}{\code{ProcessProgress$error()}}
\item \href{#method-ProcessProgress-clone}{\code{ProcessProgress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessProgress-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessProgress-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$new(label, p = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{character string}

\item{\code{p}}{progress type, logical or shiny session object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessProgress-update"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessProgress-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$update(value, msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{numeric}

\item{\code{msg}}{character string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessProgress-error"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessProgress-error}{}}}
\subsection{Method \code{error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$error(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{character string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessProgress-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessProgress-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
