% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{registerFreezedClassLabel}
\alias{registerFreezedClassLabel}
\title{Freeze Class Labels}
\usage{
registerFreezedClassLabel(output, classLabel, Results, fm, id)
}
\arguments{
\item{output}{shiny output object}

\item{classLabel}{reactive containing class labels}

\item{Results}{reactive containing the result of FutureManager}

\item{fm}{FutureManager object}

\item{id}{shiny id}
}
\description{
Freeze Class Labels
}
\details{
Used for keeping the old class labels values if Results object
is invalidated.
}
