% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{renameClassIdToLabel}
\alias{renameClassIdToLabel}
\title{Rename ClassId To Labels from classAssignment}
\usage{
renameClassIdToLabel(df, column, ca)
}
\arguments{
\item{df}{data.frame}

\item{column}{column}

\item{ca}{classAssignment object}
}
\value{
data.frame with \code{column} remapped to labels
}
\description{
Rename ClassId To Labels from classAssignment
}
\examples{

ca <- exampleClassAssigment()
setClassLabel(ca, "class1") <- "C1"
setClassLabel(ca, "class2") <- "C2"

df <- data.frame(cl = c("class1", "class2", "class1"))
renameClassIdToLabel(df, "cl", ca)

}
