% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{reorderByScore}
\alias{reorderByScore}
\title{Utility function for reordering class levels by score column}
\usage{
reorderByScore(
  df,
  orderCol = getOption("xiff.column"),
  valueCol = "score",
  .desc = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{orderCol}{column to be ordered}

\item{valueCol}{column used for ordering}

\item{.desc}{use decreasing (TRUE) or increasing order (FALSE)}
}
\value{
data.frame with \code{orderCol} transformed to ordered factor
}
\description{
Utility function for reordering class levels by score column
}
\examples{

df <- data.frame(class = c("X", "A", "C"), var = c(0,-5,5))
levels(reorderByScore(df, orderCol = "class", valueCol = "var")$class)
levels(
  reorderByScore(df, orderCol = "class", valueCol = "var", .desc = FALSE)$class
)


}
