% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_restoreSelection.R
\name{restoreSelectionInputModeUI}
\alias{restoreSelectionInputModeUI}
\alias{restoreSelectionInputMode}
\alias{saveProperties}
\title{Save Classes Properties in Shiny Application}
\usage{
restoreSelectionInputModeUI(id, ...)

restoreSelectionInputMode(input, output, session, classStack)

saveProperties(
  res,
  classSelection,
  classLabel,
  classStack,
  Annotation,
  msg = ""
)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{classStack}{reactive used to save data.frame with class properties}

\item{res}{data.frame with additional info about}

\item{classSelection}{reactive containing class selection}

\item{classLabel}{reactive containing class labels}

\item{Annotation}{annotation data}

\item{msg}{additional message to be shown when saving data}
}
\description{
Save Classes Properties in Shiny Application
}
\details{
\code{saveProperties} is a function to be used in other modules to
save the properties for \code{restoreSelectionInputMode} module.
}
