% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{runGCloudCommand}
\alias{runGCloudCommand}
\title{Run GCloud command}
\usage{
runGCloudCommand(args)
}
\arguments{
\item{args}{character vector of the command arguments, see system2 docs for
details}
}
\value{
the command output
}
\description{
Runs gcloud command with additional args using system2()
}
