% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{scaleRows}
\alias{scaleRows}
\title{Scale rows by mean and standard deviation}
\usage{
scaleRows(x)
}
\arguments{
\item{x}{matrix}
}
\value{
scaled matrix
}
\description{
Scale rows by mean and standard deviation
}
\examples{

x <- cbind(c(10,200,3000), c(20,300,5000))
scaleRows(x)

}
