% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{selectBestFeatures}
\alias{selectBestFeatures}
\alias{selectBestFeaturesTTest}
\alias{selectBestFeaturesGlmnet}
\title{Select Best Features}
\usage{
selectBestFeatures(
  df,
  threshold = "Confirmed",
  maxFeatures = Inf,
  .otherParams = list()
)

selectBestFeaturesTTest(
  df,
  threshold = 0.05,
  maxFeatures = Inf,
  .otherParams = list()
)

selectBestFeaturesGlmnet(
  df,
  threshold = 0.05,
  maxFeatures = Inf,
  .otherParams = list()
)
}
\arguments{
\item{df}{data.frame with class column and variables to be predicted}

\item{threshold}{default threshold for variable importance cutoff}

\item{maxFeatures}{max number of features to be selected}

\item{.otherParams}{other parameters to be passed into 
feature selection algorithm}
}
\value{
list with following 

\itemize{
 \item{"stats"}{data.frame with feature importance statistics}
 \item{"df"}{data.frame with the selected variables }
 \item{".otherParams"}{list of other parameters used by the feature selection
  algorithm, could be used downstream during the prediction/data preparation pahse}
 \item{"method"}{name of the feature selection metod}
}
}
\description{
Select Best Features
}
\details{
\code{selectBestFeaturesTTest} - uses the t.test

\code{selectBestFeaturesGlmnet} - uses the glmnet logistic regression.
}
\examples{

library(dplyr)
iris2 <- iris \%>\% rename(class = Species) \%>\%
  filter(class \%in\% c("setosa", "versicolor")) \%>\%
  mutate(class = droplevels(class))

selectBestFeatures(iris2)
selectBestFeaturesTTest(iris2)
selectBestFeaturesGlmnet(iris2)

}
