% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setClassItems<-}
\alias{setClassItems<-}
\title{Set items in ClassAssigment object}
\usage{
setClassItems(ca, classId) <- value
}
\arguments{
\item{ca}{classAssigment}

\item{classId}{id of class ('class1' or 'class2')}

\item{value}{character vector with new items}
}
\description{
Set items in ClassAssigment object
}
\examples{

ca <- exampleClassAssigment()
ca$class1

setClassItems(ca, "class1") <- c("A", "B")
ca$class1

}
