% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setClassLabel<-}
\alias{setClassLabel<-}
\title{Rename label in classAssignment}
\usage{
setClassLabel(ca, classId) <- value
}
\arguments{
\item{ca}{classAssignment object}

\item{classId}{which class label needs to be changed ("class1" or "class2")}

\item{value}{new label}
}
\description{
Rename label in classAssignment
}
\examples{

ca <- exampleClassAssigment()

setClassLabel(ca, "class1") <- "C1"
ca

setClassLabel(ca, "class2") <- "C2"
ca

}
