% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{setDbOptions}
\alias{setDbOptions}
\title{Set DB options}
\usage{
setDbOptions(settings = NULL)
}
\arguments{
\item{settings}{list of settings}
}
\value{
nothing
}
\description{
This function sets all R options required for DB connection based on the
settings object or environment variables (dbhost, dbname, dbuser, dbpass, dbport).
}
