% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{setPostgresql}
\alias{setPostgresql}
\title{Run set query in Postgres}
\usage{
setPostgresql(sql)
}
\arguments{
\item{sql}{character string, the SQL query}
}
\value{
TRUE if success
}
\description{
Function that queries the DB
}
