% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{withErrorHandler}
\alias{withErrorHandler}
\alias{errorScreen}
\alias{registerExtendedInputObserver}
\alias{styleHigherCol}
\alias{getEnsgRowCallback}
\alias{validateArgs}
\title{Shiny Utility Functions}
\usage{
withErrorHandler(
  expr,
  errorId = NULL,
  session = getDefaultReactiveDomain(),
  callback = NULL
)

errorScreen(reason)

registerExtendedInputObserver(
  input,
  rv,
  inputId,
  callback = NULL,
  initValue = NULL,
  debug = FALSE,
  ignoreInit = TRUE,
  ...
)

styleHigherCol(condition, cl)

getEnsgRowCallback(species, idx = 1)

validateArgs(args)
}
\arguments{
\item{expr}{}

\item{errorId}{}

\item{session}{}

\item{callback}{}
}
\description{
Shiny Utility Functions
}
