% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{shinyDropUnbalancedTumortypes}
\alias{shinyDropUnbalancedTumortypes}
\title{Shiny Drop Unbalanced Tumortypes}
\usage{
shinyDropUnbalancedTumortypes(AnnotationFocus, classSelection, minCount = 1)
}
\arguments{
\item{AnnotationFocus}{item annotation}

\item{classSelection}{classAssignment object}

\item{minCount}{min count for getBalancedTumortypes}
}
\description{
Shiny Drop Unbalanced Tumortypes
}
\details{
module for dropping unbalanced tumortypes
}
