% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{splitTrainingTestSets}
\alias{splitTrainingTestSets}
\title{Split classAssigment into training and test sets}
\usage{
splitTrainingTestSets(assignment, p_test = 0.2)
}
\arguments{
\item{assignment}{classAssigment object or data.frame}

\item{p_test}{the percentage of the test set size}
}
\value{
list with two elements - \code{training} and \code{test}
}
\description{
Split classAssigment into training and test sets
}
\examples{

ca <- exampleClassAssigment()
splitTrainingTestSets(ca)

}
