% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stackClasses}
\alias{stackClasses}
\title{Stack Class Assignment object}
\usage{
stackClasses(sampleClasses, classLabel = NULL, return_factor = FALSE)
}
\arguments{
\item{sampleClasses}{classAssignment object}

\item{classLabel}{}

\item{return_factor}{}
}
\value{
data.frame with {getOption("xiff.column")} and class columns
}
\description{
Stack Class Assignment object
}
\examples{

ca <- exampleClassAssigment()
stackClasses(ca) \%>\% str
stackClasses(ca, return_factor = TRUE) \%>\% str

}
