% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_affinity_ratio.R
\name{stripGrepModel}
\alias{stripGrepModel}
\title{Strip GREP model into simple format.}
\usage{
stripGrepModel(grepFit, geneAnno)
}
\arguments{
\item{grepFit}{GREP model.}

\item{geneAnno}{gene annotations}
}
\value{
list with modelCoefficients and epsilon
}
\description{
Strip GREP model into simple format.
}
