% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{textAreaContentToVector}
\alias{textAreaContentToVector}
\title{Convert the input of Text Area into Vector}
\usage{
textAreaContentToVector(content)
}
\arguments{
\item{content}{content of textArea}
}
\value{
character vector, or NULL if content is empty
}
\description{
Convert the input of Text Area into Vector
}
\examples{

textAreaContentToVector("")
textAreaContentToVector("A\nB\nC")

}
