% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_affinity_ratio.R
\name{tpmGREPtransform}
\alias{tpmGREPtransform}
\title{log2tpm to tpm transformation for GREP method.}
\usage{
tpmGREPtransform(x, epsilon = 0)
}
\arguments{
\item{x}{vector of log2tpm}

\item{epsilon}{value to be added after applying transformation.}
}
\value{
transformed x values.
}
\description{
log2tpm to tpm transformation for GREP method.
}
\details{
It's a little bit different than standrd transformation 2^x
but it matches the original GREP definition
}
