% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralnetworkDataClassifier.R
\docType{data}
\name{train_model_data}
\alias{train_model_data}
\title{Random data set for testing the classification models.}
\format{
A data frame with 56 rows and 7 variables:
\describe{
  \item{class}{group to classify}
  \item{other columns}{independent variables}
}
}
\source{
geneinfo2db
}
\usage{
train_model_data
}
\description{
Random data set for testing the classification models.
}
\examples{

data("train_model_data", package = "XIFF")

}
\keyword{datasets}
