% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{trainModel}
\alias{trainModel}
\title{Train Machine Learning Models.}
\usage{
trainModel(df, method = "rf", tuneLength = 5, number = 10, repeats = 10)
}
\arguments{
\item{df}{}

\item{method}{}

\item{tuneLength}{}

\item{number}{}

\item{repeats}{}
}
\description{
Train Machine Learning Models.
}
\examples{

\dontrun{
  data("train_model_data", package = "XIFF")
  train_model_data <- train_model_data[,-1]

  fitRF  <- trainModel(train_model_data, method = "rf", tuneLength = 1)
  fitSVM <- trainModel(train_model_data, method = "svmLinear2", tuneLength = 1)
  fitNeuralnet <- trainModel(train_model_data, method = "neuralnetwork")
}


}
