% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_progress.R
\name{validateFunctionResult}
\alias{validateFunctionResult}
\title{Utility function for checking the result of a function}
\usage{
validateFunctionResult(x)
}
\arguments{
\item{x}{result to be validated, if string, then the function will throw an
error.}
}
\value{
nothing, called for side effects
}
\description{
Utility function for checking the result of a function
}
