% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_utils.R
\name{validateXiffMachineLearningResult}
\alias{validateXiffMachineLearningResult}
\title{Utility function for checking if an object is XiffMachineLearningResult.}
\usage{
validateXiffMachineLearningResult(model)
}
\arguments{
\item{model}{should be XiffMachineLearningResult}
}
\value{
nothing. Called for side effect in shiny.
}
\description{
Utility function for checking if an object is XiffMachineLearningResult.
}
