% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_utils.R
\name{withHash}
\alias{withHash}
\title{Create href path css and js files}
\usage{
withHash(filePath, wwwPath)
}
\arguments{
\item{filePath}{path to css or js file}

\item{wwwPath}{href path}
}
\value{
wwwPath with md5 attached.
}
\description{
Create href path css and js files
}
\examples{

withHash(
  filePath = system.file("www/common.css", package = "XIFF"),
  wwwPath = "xiff/common.css"
)

}
