% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generatePlotByType}
\alias{generatePlotByType}
\alias{generateROCPlot}
\alias{generateDiffPlot}
\title{Basic Plots}
\usage{
generatePlotByType(
  data,
  plotType,
  dataCol,
  title = NULL,
  ca = NULL,
  rocPlotFun = generateROCPlot,
  diffPlotFun = generateDiffPlot,
  ...
)

generateROCPlot(data, ca, dataCol, title = "ROC plot")

generateDiffPlot(
  data,
  ca,
  dataCol,
  plotFunc,
  title = NULL,
  xlabel = "",
  ylabel = "",
  trans = "identity",
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{plotType}{type of the plot (all, point, roc, violin, box, coverage)}

\item{dataCol}{which column}

\item{title}{ggplot2 title}

\item{ca}{classAssigment object}

\item{rocPlotFun}{function used to create ROC plots}

\item{diffPlotFun}{function used to make diffPlotFun}

\item{...}{other arguments passed to the plotting function}
}
\value{

}
\description{
Basic Plots
}
\examples{

df <- exampleDataFrame()
ca <- exampleClassAssigment()
generatePlotByType(df, "point", "tpm", "Plot", ca = ca)
generatePlotByType(df, "roc", "tpm", "Plot", ca = ca)

}
