% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{xiffSupportedModels}
\alias{xiffSupportedModels}
\title{Get a vector machine learning models that are supported by XIFF package.}
\usage{
xiffSupportedModels()
}
\value{
a character vector containing the model names that are supported by XIFF package.
}
\description{
Get a vector machine learning models that are supported by XIFF package.
}
\details{
First element should be treated as default model.
}
