% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{appUI_main_about}
\alias{appUI_main_about}
\alias{appUI_title_right}
\alias{appUI_title}
\title{Tabsets for shiny application}
\usage{
appUI_main_about(id, docuLink, aboutTabUIFunc)

appUI_title_right(id, docuLink, packageName, dbName)

appUI_title(id, title, logoPath)
}
\arguments{
\item{id}{shiny id}

\item{docuLink}{link to the documentation}

\item{aboutTabUIFunc}{function that creates about ui}
}
\description{
Tabsets for shiny application
}
