% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{appTheme}
\alias{appTheme}
\title{Default app theme}
\usage{
appTheme(version = 4, font = "Roboto", ...)
}
\arguments{
\item{version}{bootstrap version to use}

\item{font}{font name to use}

\item{...}{arguments passed to bs_theme}
}
\value{

}
\description{
bslib theme to be used in apps
}
