% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{columnTabPanel}
\alias{columnTabPanel}
\title{Column Tab Panel}
\usage{
columnTabPanel(title, value, inputMenu, outputArea)
}
\arguments{
\item{title}{character string}

\item{value}{character string}

\item{inputMenu}{shiny tag}

\item{outputArea}{shiny tag}
}
\description{
Column Tab Panel
}
