% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{createMachineLearningModel}
\alias{createMachineLearningModel}
\title{Create machine learning model using XIFF package.}
\usage{
createMachineLearningModel(
  trainingSet,
  geneSet,
  geneAnno,
  p = FALSE,
  classLabel = list(class1_name = "class1", class2_name = "class2"),
  method = "rf",
  tuneLength = 5,
  number = 10,
  repeats = 10,
  selectBestFeaturesFnc = "auto",
  threshold = "Confirmed",
  maxFeatures = "auto",
  ...,
  .verbose = TRUE
)
}
\description{
Create machine learning model using XIFF package.
}
