% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateDataCoveragePlot}
\alias{generateDataCoveragePlot}
\title{Generate Data Coverage Plot}
\usage{
generateDataCoveragePlot(data, col, ca, addCountLabels = TRUE)
}
\arguments{
\item{data}{data.frame with data.}

\item{col}{column from \code{data} to be used in creating the plot}

\item{ca}{classAssignment object}

\item{addCountLabels}{if TRUE (default),
add labels with available observations}
}
\description{
Generate Data Coverage Plot
}
\examples{

library(dplyr)
library(tibble) 
ca <- classAssignment(a = c(1:10), b = c(11:25))

dt <- tibble(celllinename = as.character(1:25), value = rep("data", 25))
dt[c(2,3,11:16),"value"] <- NA

generateDataCoveragePlot(dt, "value", ca)
generateDataCoveragePlot(dt, "value", ca, addCountLabels = FALSE)

}
