% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getExpressionDimRedPlot}
\alias{getExpressionDimRedPlot}
\alias{generateExpressionDimRedPlot}
\title{generateExpressionDimRedPlot}
\usage{
getExpressionDimRedPlot(
  dimRedResult,
  colorCol = "class",
  labelCol = NULL,
  fontSize = 10
)

generateExpressionDimRedPlot(
  dimRedResult,
  colorCol = "class",
  labelCol = NULL,
  fontSize = 10
)
}
\arguments{
\item{dimRedResult}{result of \code{getExpressionDimRed}}

\item{colorCol}{column name for coloring.}

\item{labelCol}{column name for labels.}

\item{fontSize}{font size for labels}
}
\description{
generateExpressionDimRedPlot
}
