% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateGSEA_plot}
\alias{generateGSEA_plot}
\title{Generate GSEA Plot}
\usage{
generateGSEA_plot(diffExResult, ensg_geneset, rankType, classLabels = NULL)
}
\arguments{
\item{diffExResult}{}

\item{ensg_geneset}{}

\item{rankType}{"p.valueDir", "logFC"}

\item{classLabels}{}
}
\value{
ggplot2 plot
}
\description{
Generate GSEA Plot
}
\examples{

if(require("CLIFF")) {
   gene_anno <- CLIFF::getGeneAnno("human")
   cs <- CLIFF::exampleClassAssigment()
   limmaData <- CLIFF::differentialGeneExpression_LimmaVoom(cs, gene_anno)
   geneset <- CLIFF::getGSEAdata("human", gene_set = "HALLMARK_P53_PATHWAY")
   
   generateGSEA_plot(
     limmaData,
     ensg_geneset = geneset,
     rankType = "p.valueDir"
   )
}

}
