% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_plots.R
\name{generateTablePlot}
\alias{generateTablePlot}
\title{Generate Visualization of the confusion matrix.}
\usage{
generateTablePlot(df)
}
\arguments{
\item{df}{MLModelTestsResult or data.frame with columns predicted and reference columns}
}
\value{
ggplot object
}
\description{
Generate Visualization of the confusion matrix.
}
\examples{

tbl <- structure(
  list(
    predicted = structure(
      c(1L, 1L, 1L, 1L, 1L, 1L, 2L, 2L),
      .Label = c("resistant\n(negative)", "sensitive\n(positive)"),
      class = "factor"
    ),
    reference = structure(
      c(1L, 2L, 2L, 2L,
        2L, 2L, 2L, 1L),
      .Label = c("sensitive\n(positive)", "resistant\n(negative)"),
      class = "factor"
    )
  ),
  row.names = c(NA,-8L),
  class = c("tbl_df",
            "tbl", "data.frame")
)
generateTablePlot(tbl)

}
