% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_bioinfo.R
\name{getMetadataGeneSetTable}
\alias{getMetadataGeneSetTable}
\alias{getGeneSet}
\title{Get Gene Set Information}
\usage{
getMetadataGeneSetTable()

getGeneSet(geneset)
}
\arguments{
\item{geneset}{gene set name}
}
\value{
vector with genes ensg
}
\description{
Get Gene Set Information
}
\examples{

setDbOptions()
meta <- getMetadataGeneSetTable()
meta

head(getGeneSet(meta$genesetname[1]))

}
