% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_bioinfo.R
\name{prepareClassificationTable}
\alias{prepareClassificationTable}
\title{Prepare Classification Table}
\usage{
prepareClassificationTable(class1, class2, addRownames = TRUE)
}
\arguments{
\item{class1}{vector with class one items}

\item{class2}{vector with class two items}

\item{addRownames}{logical, if \code{TRUE} then adds items as rownames}
}
\value{
data.frame
}
\description{
Prepare Classification Table
}
\examples{

prepareClassificationTable(c("A", "B"), c("C", "D"))

}
