% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_progress.R
\name{ProcessProgress}
\alias{ProcessProgress}
\title{R6 ProcessProgress bar}
\description{
R6 ProcessProgress bar

R6 ProcessProgress bar
}
\details{
Progress bar for shiny
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProcessProgress$new()}}
\item \href{#method-update}{\code{ProcessProgress$update()}}
\item \href{#method-error}{\code{ProcessProgress$error()}}
\item \href{#method-clone}{\code{ProcessProgress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$new(label, p = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$update(value, msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-error"></a>}}
\if{latex}{\out{\hypertarget{method-error}{}}}
\subsection{Method \code{error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$error(msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessProgress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
