% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_programmerImport.R
\name{programmerImportInputModeUI}
\alias{programmerImportInputModeUI}
\alias{programmerImportInputMode}
\title{Programmer Import Input Module}
\usage{
programmerImportInputModeUI(id)

programmerImportInputMode(input, output, session, Annotation)
}
\arguments{
\item{id}{module id}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{Annotation}{annotation data.frame}
}
\description{
Programmer Import Input Module
}
