% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tabLayout.R
\name{tabLayoutUI_main}
\alias{tabLayoutUI_main}
\alias{tabLayoutUI_sidebar}
\alias{tabLayout}
\title{Tab module layout}
\usage{
tabLayoutUI_main(id, useMiddlePlot = TRUE)

tabLayoutUI_sidebar(
  id,
  defaults = list(),
  input = list(),
  additionalChoices = c(),
  hidden = c(),
  useMiddlePlot = TRUE
)

tabLayout(
  input,
  output,
  session,
  plotFun,
  TableData,
  jsRowCallback = htmlwidgets::JS("preventLinkSelections"),
  otherDToptions = list()
)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{plotFun}{}

\item{TableData}{}

\item{jsRowCallback}{}
}
\description{
Tab module layout
}
