% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_upload.R
\name{uploadInputModeUI}
\alias{uploadInputModeUI}
\alias{uploadInputMode}
\title{Upload Input Mode}
\usage{
uploadInputModeUI(id, allowRds = FALSE, helpUrl = NULL)

uploadInputMode(
  input,
  output,
  session,
  AnnotationFull,
  translationFun,
  AllTumortype = NULL,
  AnnotationFiltered = NULL,
  mlUseTumortypeFilter = TRUE
)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{AnnotationFull}{}

\item{translationFun}{}

\item{AllTumortype}{}

\item{AnnotationFiltered}{}

\item{mlUseTumortypeFilter}{}
}
\description{
Upload Input Mode
}
